MAINT: undo the assumption of d=2
This evolves into a mess. This probably stems from the requirement that
index_type has always the correct length. The simplest way to ensure that
is to have a constructor from num_dim, but Cython does not seem
to allow stack-allocated objects with non-default constructors. Fun.

Likely this all will need to be reworked to
(i) hide map_type::index_type from Cython (it does not understand
    nested typedefs anyway)
(ii) use plain C arrays with MAX_DIMS or smth like that